+++ command/part.go
@@ -15,6 +15,7 @@ var part = Command{
 	Description: "Leave the message author's channel",
 	Cooldown:    5,
 	NoPrefix:    false,
+	CanDisable:  false,
 	Execute: func(message *Message, sender MessageSender, args []string) error {
 		tx, err := message.DB.Begin()
 		if err != nil {