+++ command/ping.go
@@ -12,6 +12,7 @@ var ping = Command{
 	Description: "Responds with pong and latency to twitch in milliseconds",
 	Cooldown:    5,
 	NoPrefix:    false,
+	CanDisable:  false,
 	Execute: func(message *Message, sender MessageSender, args []string) error {
 		latency := fmt.Sprintf("%d ms", time.Since(message.Time).Milliseconds())
 		sender.Say(message.Channel, fmt.Sprintf("🐒 Pong! Latency: %s", latency))